(refactor) computation jitter coords
Jitter for center patch coordinates was computed on world coordinates instead of voxel coordinates. Not wrong, but this method ensures the max jitter is computed on voxel diameter instead of length in mm.
Also deleted unused function used to compute if coordinate was close to aneurism. Already execute a faster computation in patch_creation script.